Edit lefigaro.fr rule (read description)
For this website, it’s impossible to completely avoid the unintended effects of the cookie banner. When you click “reject cookies,” you are redirected to a subscription page. Therefore, to access the website without cookies and without a subscription, you need to go back once you’re on the subscription page.